Config docs and `index` entry

+++ singlet/io/pickle/__init__.py
@@ -3,10 +3,10 @@
 # date:       02/08/17
 # content:    Support module for filenames related to pickle files.
 # Modules
+import pandas as pd
 
 
 # Parser
 def parse_counts_table(path, fmt):
-    import pandas as pd
-    table = pd.read_pickle(path).astype(float)
+    table = pd.read_pickle(path)
     return table
